requirements.txt
discord
flask
python-dotenv
Flask-SQLAlchemy